refactor(ui): address catalog settings review nits
Drop shared barrel exports, import modules directly, remove validation
pass-through re-exports, and document mount-only auto-preview effect.
