+++ clients/ui/frontend/src/app/api/mcpCatalogSettings/service.ts
@@ -5,7 +5,7 @@ import {
   McpCatalogSourcePreviewRequest,
   McpCatalogSourcePreviewResult,
 } from '~/app/mcpServerCatalogTypes';
-import { createSourceConfigService } from '~/app/shared/catalogSettings';
+import { createSourceConfigService } from '~/app/shared/catalogSettings/api/createSourceConfigService';
 
 const service = createSourceConfigService<
   McpCatalogSourceConfig,
