+++ clients/ui/frontend/src/app/context/mcpCatalogSettings/McpCatalogSettingsContext.tsx
@@ -8,7 +8,7 @@ import type { McpCatalogSourceConfigList } from '~/app/mcpServerCatalogTypes';
 import type { CatalogSourceList } from '~/app/shared/types/catalogTypes';
 import { BFF_API_VERSION, URL_PREFIX } from '~/app/utilities/const';
 import useModelCatalogAPIState from '~/app/hooks/modelCatalog/useModelCatalogAPIState';
-import { useCatalogSourcesWithPolling } from '~/app/shared/catalogSettings';
+import { useCatalogSourcesWithPolling } from '~/app/shared/catalogSettings/hooks/useCatalogSourcesWithPolling';
 
 export type McpCatalogSettingsContextType = {
   apiState: McpCatalogSettingsAPIState;
