+++ clients/ui/frontend/src/app/hooks/mcpCatalogSettings/useMcpCatalogSourceConfigBySourceId.ts
@@ -2,7 +2,7 @@ import { FetchState } from 'mod-arch-core';
 import * as React from 'react';
 import { McpCatalogSourceConfig } from '~/app/mcpServerCatalogTypes';
 import { McpCatalogSettingsContext } from '~/app/context/mcpCatalogSettings/McpCatalogSettingsContext';
-import { useSourceConfigById } from '~/app/shared/catalogSettings';
+import { useSourceConfigById } from '~/app/shared/catalogSettings/hooks/useSourceConfigById';
 
 export const useMcpCatalogSourceConfigBySourceId = (
   sourceId: string,
