+++ clients/ui/frontend/src/app/hooks/mcpCatalogSettings/useMcpCatalogSourceConfigs.ts
@@ -1,6 +1,6 @@
 import { FetchState } from 'mod-arch-core';
 import { McpCatalogSourceConfigList } from '~/app/mcpServerCatalogTypes';
-import { useSourceConfigs } from '~/app/shared/catalogSettings';
+import { useSourceConfigs } from '~/app/shared/catalogSettings/hooks/useSourceConfigs';
 import { McpCatalogSettingsAPIState } from './useMcpCatalogSettingsAPIState';
 
 export const useMcpCatalogSourceConfigs = (
