+++ clients/ui/frontend/src/app/hooks/modelCatalogSettings/useCatalogSourceConfigBySourceId.ts
@@ -2,7 +2,7 @@ import { FetchState } from 'mod-arch-core';
 import * as React from 'react';
 import { CatalogSourceConfig } from '~/app/modelCatalogTypes';
 import { ModelCatalogSettingsContext } from '~/app/context/modelCatalogSettings/ModelCatalogSettingsContext';
-import { useSourceConfigById } from '~/app/shared/catalogSettings';
+import { useSourceConfigById } from '~/app/shared/catalogSettings/hooks/useSourceConfigById';
 
 export const useCatalogSourceConfigBySourceId = (
   sourceId: string,
