+++ clients/ui/frontend/src/app/pages/mcpCatalogSettings/utils/validation.ts
@@ -1,28 +1,23 @@
 import { ManageMcpSourceFormData } from '~/app/pages/mcpCatalogSettings/useManageMcpSourceData';
+import { SOURCE_NAME_CHARACTER_LIMIT } from '~/app/shared/catalogSettings/const';
 import {
-  isSourceNameEmpty,
-  SOURCE_NAME_CHARACTER_LIMIT,
   validateSourceName as validateSharedSourceName,
   validateYamlContent,
-} from '~/app/shared/catalogSettings';
+} from '~/app/shared/catalogSettings/utils/validation';
 
 export const validateMcpSourceName = (name: string): boolean =>
   validateSharedSourceName(name, SOURCE_NAME_CHARACTER_LIMIT);
-
-export const isMcpSourceNameEmpty = isSourceNameEmpty;
-
-export const validateMcpYamlContent = validateYamlContent;
 
 export const isMcpFormValid = (data: ManageMcpSourceFormData): boolean => {
   if (data.isDefault) {
     return true;
   }
-  return validateMcpSourceName(data.name) && validateMcpYamlContent(data.yamlContent);
+  return validateMcpSourceName(data.name) && validateYamlContent(data.yamlContent);
 };
 
 export const isMcpPreviewReady = (data: ManageMcpSourceFormData): boolean => {
   if (data.isDefault) {
     return true;
   }
-  return validateMcpYamlContent(data.yamlContent);
+  return validateYamlContent(data.yamlContent);
 };
