+++ clients/ui/frontend/src/app/pages/modelCatalogSettings/components/PreviewPanel.tsx
@@ -28,7 +28,7 @@ import {
   UseSourcePreviewResult,
   PreviewMode,
 } from '~/app/pages/modelCatalogSettings/useSourcePreview';
-import { CatalogSettingsPreviewTab } from '~/app/shared/catalogSettings';
+import { CatalogSettingsPreviewTab } from '~/app/shared/catalogSettings/hooks/previewTypes';
 import PreviewButton from './PreviewButton';
 
 type PreviewPanelProps = {
