--- clients/ui/frontend/src/app/shared/catalogSettings/index.ts
@@ -1,17 +1,0 @@
-export { SOURCE_NAME_CHARACTER_LIMIT } from './const';
-export { generateSourceIdFromName } from './utils/generateSourceIdFromName';
-export { parseCommaSeparatedList } from './utils/parseCommaSeparatedList';
-export {
-  isNonEmptyString,
-  validateSourceName,
-  isSourceNameEmpty,
-  validateYamlContent,
-} from './utils/validation';
-
-export { createSourceConfigService } from './api/createSourceConfigService';
-
-export { useCatalogSourcesWithPolling } from './hooks/useCatalogSourcesWithPolling';
-export { useSourceConfigs } from './hooks/useSourceConfigs';
-export { useSourceConfigById } from './hooks/useSourceConfigById';
-export { useCatalogSourcePreviewCore } from './hooks/useCatalogSourcePreviewCore';
-export { CatalogSettingsPreviewTab } from './hooks/previewTypes';
